Fix remaining utils imports after moving to lib directory
Update utils imports in pages/index.js and test files to use
the new lib/utils.js location.
